+++ crates/compiler/src/driver_constants.rs
@@ -33,7 +33,7 @@ pub mod addresses {
         EON_SHADOW_MUSIC,
         EON_SHADOW_SFX,
         SONG_TICK_COUNTER,
-        ENABLED_CHANNELS_MASK,
+        IO_MUSIC_CHANNELS_MASK,
         CHANNEL_VC_VOL_L,
         CHANNEL_VC_VOL_R,
         CHANNEL_VC_PITCH_L,